cover with tests scrollViewDelegate
